kernel: fix x86_64 threading, PIC/PIT, and context restore
- Fix timeout sentinel check in lock_mutex/wait_cond: use u32::MAX
  instead of usize::MAX since kcall args are u32-width
- Enable PIC and PIT port registration for x86_64 (was skipped,
  causing interrupt controller init to fail)
- Fix context_restore in hooks.S: restore segment registers before
  GPRs to avoid clobbering RAX with the GS selector value

All thread-rust tests pass (threads, mutexes, condvars, timers,
TDA, scheduler) - 5/5 stable runs.

+++ src/kernel/src/pm/kcall/lock_mutex.rs
@@ -73,7 +73,8 @@ pub unsafe fn lock_mutex(
     );
     // Unpack kernel call arguments.
     let mutex_addr: MutexAddress = MutexAddress::from(mutex_addr);
-    let timeout: Option<SystemTime> = if timeout_s == usize::MAX && timeout_ns == usize::MAX {
+    // Use u32::MAX as the sentinel because kcall arguments are u32-width.
+    let timeout: Option<SystemTime> = if timeout_s == u32::MAX as usize && timeout_ns == u32::MAX as usize {
         None
     } else {
         match SystemTime::new(timeout_s as u64, timeout_ns as u32) {
